refactor core as kv_manager
